pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools"

[project]
name = "manga-video-pipeline"
version = "0.1.0"
description = "An automated manga video pipeline for creating content from trending manga"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Rowan Lee", email = "rowan@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "playwright>=1.44.0",
    "beautifulsoup4>=4.12.3",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "ffmpeg-python>=0.2.0",
    "pillow>=10.3.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "celery>=5.4.0",
    "redis>=5.0.4",
    "google-api-python-client>=2.130.0",
    "google-auth-oauthlib>=1.2.0",
    "requests-oauthlib>=2.0.0",
    "python-telegram-bot>=21.2",
    "pydantic-settings>=2.3.0",
    "structlog>=24.2.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pytest-cov"
]

[project.scripts]
manga-pipeline = "manga_video_pipeline.cli:main"

[project.urls]
Homepage = "https://github.com/rowanlee/manga-video-pipeline"
Repository = "https://github.com/rowanlee/manga-video-pipeline"